fix library panel differences between backend and frontend conversion
