+++ public/app/features/dashboard/api/ResponseTransformersToBackend.test.ts
@@ -1,13 +1,9 @@
 import { readdirSync, readFileSync } from 'fs';
 import path from 'path';
 
-import {
-  Spec as DashboardV2Spec,
-  GridLayoutItemKind,
-  RowsLayoutRowKind,
-} from '@grafana/schema/dist/esm/schema/dashboard/v2';
 import { mockDataSource } from 'app/features/alerting/unified/mocks';
 import { setupDataSources } from 'app/features/alerting/unified/testSetup/datasources';
+import { normalizeBackendOutputForFrontendComparison } from 'app/features/dashboard-scene/serialization/serialization-test-utils';
 import { MIXED_DATASOURCE_NAME } from 'app/plugins/datasource/mixed/MixedDataSource';
 
 import { ensureV2Response } from './ResponseTransformers';
@@ -184,71 +180,6 @@ describe('Backend / Frontend result comparison', () => {
 
   // Filter to only process v1beta1 input files
   const v1beta1Inputs = jsonInputs.filter((inputFile) => inputFile.startsWith('v1beta1.'));
-
-  /**
-   * Normalizes backend output to match frontend behavior.
-   * The backend sets repeat properties on library panel grid items from the library panel definition,
-   * but the frontend only sets repeat when explicitly set on the panel instance.
-   * This function removes repeat properties from library panel items where they weren't set on the instance.
-   *
-   * The difference in behavior is due to how the frontend conversion is done.
-   * It is not feasible to fetch all library panels async in all cases where the transformation is done.
-   * Library panel repeats will be set by the library panel behavior in those cases.
-   */
-  function normalizeBackendOutputForFrontendComparison(
-    backendSpec: DashboardV2Spec,
-    inputPanels: Array<{ id?: number; libraryPanel?: { uid?: string }; repeat?: string }>
-  ): DashboardV2Spec {
-    const normalized = JSON.parse(JSON.stringify(backendSpec)) as DashboardV2Spec;
-
-    // Create a map of panel ID to whether it has explicit repeat
-    const panelHasExplicitRepeat = new Map<number, boolean>();
-    inputPanels.forEach((panel) => {
-      if (panel.id !== undefined) {
-        panelHasExplicitRepeat.set(panel.id, !!panel.repeat);
-      }
-    });
-
-    // Helper to recursively process grid items
-    function processGridItems(items: GridLayoutItemKind[]): void {
-      if (!Array.isArray(items)) {
-        return;
-      }
-
-      items.forEach((item) => {
-        if (item.spec?.element?.name) {
-          // Extract panel ID from element name (format: "panel-{id}")
-          const match = item.spec.element.name.match(/^panel-(\d+)$/);
-          if (match) {
-            const panelId = parseInt(match[1], 10);
-            const hasExplicitRepeat = panelHasExplicitRepeat.get(panelId);
-
-            // If this is a library panel item and repeat wasn't explicitly set on the instance,
-            // remove the repeat property (backend adds it from library panel definition)
-            if (hasExplicitRepeat === false && item.spec.repeat) {
-              delete item.spec.repeat;
-            }
-          }
-        }
-      });
-    }
-
-    // Process GridLayout items
-    if (normalized.layout?.kind === 'GridLayout' && normalized.layout.spec?.items) {
-      processGridItems(normalized.layout.spec.items);
-    }
-
-    // Process RowsLayout items
-    if (normalized.layout?.kind === 'RowsLayout' && normalized.layout.spec?.rows) {
-      normalized.layout.spec.rows.forEach((row: RowsLayoutRowKind) => {
-        if (row.spec?.layout?.kind === 'GridLayout' && row.spec.layout.spec?.items) {
-          processGridItems(row.spec.layout.spec.items);
-        }
-      });
-    }
-
-    return normalized;
-  }
 
   v1beta1Inputs.forEach((inputFile) => {
     it(`should convert ${inputFile} spec to match backend conversion`, async () => {
